apps/phase4-google-adk/pyproject.toml
[project]
name = "phase4-google-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.28.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
